+++ src/lib.rs
@@ -98,7 +98,6 @@ These features enable the detection of file formats based on other ones by readi
   is not recognized by its signature.
 - `reader-xml` - Enables [Extensible Markup Language (XML)](`FileFormat::ExtensibleMarkupLanguage`)
   based file formats detection:
-  * [DRAWIO](`FileFormat::Drawio`)
   * [Extensible 3D Graphics (X3D)](`FileFormat::Extensible3DGraphics`)
   * [Extensible Stylesheet Language Transformations (XSLT)](`FileFormat::ExtensibleStylesheetLanguageTransformations`)
   * [GPS Exchange Format (GPX)](`FileFormat::GpsExchangeFormat`)
@@ -109,6 +108,7 @@ These features enable the detection of file formats based on other ones by readi
   * [Scalable Vector Graphics (SVG)](`FileFormat::ScalableVectorGraphics`)
   * [Simple Object Access Protocol (SOAP)](`FileFormat::SimpleObjectAccessProtocol`)
   * [XML Localization Interchange File Format (XLIFF)](`FileFormat::XmlLocalizationInterchangeFileFormat`)
+  * [draw.io (DRAWIO)](`FileFormat::Drawio`)
 - `reader-zip` - Enables [ZIP](`FileFormat::Zip`)-based file formats detection:
   * [3D Manufacturing Format (3MF)](`FileFormat::ThreeDimensionalManufacturingFormat`)
   * [Android Package (APK)](`FileFormat::AndroidPackage`)