+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - DRAWIO
 - InDesign Markup Language (IDML)
+- MS-DOS Batch (Batch)
 - Microsoft Visual Studio Solution (SLN)
 
 # Version 0.14.0 (2023-01-30)