+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Version 0.15.0 (unreleased)
+# Version 0.15.0 (2023-03-06)
 
 ## Improvements